+++ General/CommandLine.vb
@@ -98,7 +98,6 @@ Namespace CommandLine
         Property NoSwitch As String
         Property Path As String
         Property Switch As String
-        Property Switch2 As String
         Property Switches As IEnumerable(Of String)
         Property Text As String
         Property URLs As List(Of String)
@@ -117,7 +116,6 @@ Namespace CommandLine
             Dim ret As New HashSet(Of String)
 
             If Switch <> "" Then ret.Add(Switch)
-            If Switch2 <> "" Then ret.Add(Switch2)
             If NoSwitch <> "" Then ret.Add(NoSwitch)
             If HelpSwitch <> "" Then ret.Add(HelpSwitch)