+++ General/PowerShell.vb
@@ -4,6 +4,7 @@ Imports System.Management.Automation
 Imports System.Management.Automation.Runspaces
 
 Public Class PowerShell
+    Shared References As New List(Of Object)
     Shared Property InitCode As String
 
     Shared Function Invoke(code As String) As Object
@@ -11,29 +12,36 @@ Public Class PowerShell
     End Function
 
     Shared Function Invoke(code As String, varName As String, varValue As Object) As Collection(Of PSObject)
-        Using runspace = RunspaceFactory.CreateRunspace()
-            runspace.ApartmentState = Threading.ApartmentState.STA
-            runspace.Open()
+        Dim runspace = RunspaceFactory.CreateRunspace()
+        runspace.ApartmentState = Threading.ApartmentState.STA
+        runspace.Open()
 
-            Using pipeline = runspace.CreatePipeline()
-                Dim cmd = New Runspaces.Command("Set-ExecutionPolicy")
-                cmd.Parameters.Add("ExecutionPolicy", "Unrestricted")
-                cmd.Parameters.Add("Scope", "Process")
-                pipeline.Commands.Add(cmd)
+        Dim pipeline = runspace.CreatePipeline()
+        Dim cmd = New Runspaces.Command("Set-ExecutionPolicy")
+        cmd.Parameters.Add("ExecutionPolicy", "Unrestricted")
+        cmd.Parameters.Add("Scope", "Process")
+        pipeline.Commands.Add(cmd)
 
-                If InitCode <> "" Then
-                    pipeline.Commands.AddScript(InitCode)
-                End If
+        If InitCode <> "" Then
+            pipeline.Commands.AddScript(InitCode)
+        End If
 
-                pipeline.Commands.AddScript(code)
+        pipeline.Commands.AddScript(code)
 
-                If varName <> "" Then
-                    runspace.SessionStateProxy.SetVariable(varName, varValue)
-                End If
+        If varName <> "" Then
+            runspace.SessionStateProxy.SetVariable(varName, varValue)
+        End If
 
-                Return pipeline.Invoke()
-            End Using
-        End Using
+        If code.Contains("Register-ObjectEvent") Then
+            References.Add(runspace)
+            References.Add(pipeline)
+            Return pipeline.Invoke()
+        Else
+            Dim ret = pipeline.Invoke()
+            pipeline.Dispose()
+            runspace.Dispose()
+            Return ret
+        End If
     End Function
 
     Shared Function InvokeAndConvert(code As String) As String
@@ -56,10 +64,5 @@ Public Class PowerShell
         Next
 
         Return String.Join(Environment.NewLine, lines)
-    End Function
-
-    Shared Function ConvertToCSV(delimiter As String, objects As IEnumerable(Of Object)) As String
-        Dim code = $"$inputVar | ConvertTo-Csv -Delimiter '{delimiter}' -NoTypeInformation"
-        Return $"sep={delimiter}" + Environment.NewLine + InvokeAndConvert(code, "inputVar", objects)
     End Function
 End Class