+++ Changelog.md
@@ -2,10 +2,11 @@
 2.1.1.9 Beta (not yet available)
 ============
 
-- the cut feature was missing the last frame.
-- in case the cut feature is used, flac input for qaac is disabled
+- The cut feature was missing the last frame.
+- In case the cut feature is used, flac input for qaac is disabled
   so if necessary a w64 file is created.
-- many nvenc improvements
+- Many nvenc improvements.
+- PowerShell script host supports events, script examples use events.
 - mvtools2 2.7.43